reorganize data input
