+++ src/main/resources/input/2020/eight_instruction-list.txt
