+++ src/main/resources/input/2020/four_passport-list.txt
