+++ src/main/resources/input/2020/seven_regulation-list.txt
