+++ src/main/resources/input/2020/six_customs-list.txt
