+++ src/main/resources/input/2020/three_tree-map.txt
